Fix "global" next button behavior
Currently, ModalWizard reports an error if any form in the entire
wizard is invalid. This introduces a bug where, if a user is on a step
with a form error, and then clicks the back button, the next button
becomes permanently disabled (even if the current form is valid)

To prevent this, we also check to make sure that we are on the proper
step before disabling the button.

Included in this commit
* Disable the next button on the MappingWizardClustersStep if there are
no mappings
* Fix the global error behavior
* Remove lodash as it is no longer being used